requirements.txt
# ─── AI / LLM ────────────────────────────────
ollama>=0.4.0

# ─── Memory / RAG ────────────────────────────
chromadb>=0.5.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# ─── Document Loaders ────────────────────────
pypdf>=4.0.0

# ─── CLI / Output ────────────────────────────
rich>=13.0.0

# ─── Config ──────────────────────────────────
pyyaml>=6.0
python-dotenv>=1.0.0

# ─── Gateway Server ──────────────────────────
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.0
websockets>=14.0

# ─── Communication ───────────────────────────
python-telegram-bot>=21.0

# ─── Scheduler ───────────────────────────────
apscheduler>=3.10.0

# ─── Skills System ───────────────────────────
watchdog>=6.0.0
duckduckgo-search>=7.0.0
psutil>=6.0.0
